♻️ Refacto: candidature projector
